Decoder improvements and test program (#6017)
* Add some extra error codes

End-of-file is useful to distinguish as an error code to avoid needing
a separate way of signalling it along with success.

External libraries can fail in ways we can't easily turn into a normal
error code, but we'd still like to distinguish that from completely
unknown failure.

* Improve hardware decoder support

Add support for VAAPI on Linux, and improve how all platforms work when
testing different hardware codecs.  Also remove vestigial libmfx support
(it's just a wrapper around the normal platform API, so it never adds
anything for us).

* Include FFmpeg logs with Whist logs

Adds a command-line option to set the level of FFmpeg logging we want
to see, defaulting to the WARNING level.

* Fix interference between command-line and SDL tests

If the command-line test runs first then the global variable setting
breaks the SDL test on Windows and Linux.

* Reduce duplication in copying runtime libraries

* Remove libmfx headers and libraries

We're not going to use these - decode is handled by D3D11/VAAPI, encode
is not happened on Intel hardware.

* Add decoder test program

This allows standalone testing all of our decode path and some of
rendering using input from a file.

+++ protocol/test/codec_test.cpp
@@ -129,7 +129,7 @@ TEST_F(CodecTest, DecodeTest) {
     for (int n = 0; n < 10; n++) {
         const DecodeTestInput *input = &decode_test_input[n];
 
-        ret = video_decoder_send_packets(dec, (void *)input->packet, (int)input->size);
+        ret = video_decoder_send_packets(dec, (void *)input->packet, input->size);
         EXPECT_EQ(ret, 0);
 
         ret = video_decoder_decode_frame(dec);
